+++ src/Exceptions/InvalidControllerException.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace JTL\Shop5Router\Exceptions;
+
+/**
+ * Class InvalidControllerException
+ * @package JTL\Shop5Router\Exceptions
+ */
+class InvalidControllerException extends \Exception
+{
+}